Network/pyproject.toml
[project]
name = "gateway-network"
version = "0.1.0"
description = "GatewayIDE Network Service (WireGuard control-plane API)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "loguru>=0.7",
]

[tool.uv]
# optional, nice-to-have; defaults are fine
